Split registering Controllers, Factories and Listneres to seprate files from state middleware
